rudimentary boot tasks
It will probably be a while before we release this as an OSS project, but for the time being, we can at least start using it in scripts.

`boot pom jar install` will install adzerk-clj-0.0.1.jar to your local Maven repo, so you can then use it in Boot scripts. I'm thinking we can bump the version to 0.1.0 once we're at the point where the repo is public (and maybe moved to the adzerk-oss org) and we start deploying it to Clojars.